Update partitions.sh
When this script is run against an Oracle Virtual Box (Configured by Vagrant) this script yields a value of 0 (CPU and CORE expressions both yield 1).
This is not acceptable as it causes the reliant process to fail. 0 or 1 are not a valid number when partitioning a table. 
Default to 2 where this occurs.